% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_mds}
\alias{get_mds}
\title{get_mds}
\usage{
get_mds(...)
}
\arguments{
\item{...}{pass function arguments to plotMDS.}
}
\value{
a list with information about MDS, including plotting data.
}
\description{
Get plotting data from plotMDS() in the limma package
}
\examples{
NULL
}
