% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getEnrichmentResults}
\alias{getEnrichmentResults}
\title{Create a matrix with several objects resulting from enrichment analysis with kegga() and goana()}
\usage{
getEnrichmentResults(..., what = "P.Up", use.name = FALSE)
}
\arguments{
\item{...}{results from enrichment}

\item{what}{what to plot: P.Up (default) or P.Down.}

\item{use.name}{logical; whether to return the term's name or the ids.}
}
\description{
Create a matrix with several objects resulting from enrichment analysis with kegga() and goana()
}

