% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getKEGGResults}
\alias{getKEGGResults}
\title{Create a matrix with several objects resulting from enrichment analysis with kegga() and goana()}
\usage{
getKEGGResults(..., what = "P.Up")
}
\arguments{
\item{...}{results from enrichment.}

\item{what}{what to plot: P.Up (default) or P.Down.}
}
\description{
Create a matrix with several objects resulting from enrichment analysis with kegga() and goana()
}

