% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getMDS}
\alias{getMDS}
\title{Get plotting data from plotMDS() in the limma package}
\usage{
getMDS(...)
}
\arguments{
\item{...}{pass function arguments to plotMDS.}
}
\description{
Get plotting data from plotMDS() in the limma package
}

