% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{imputeGroup}
\alias{imputeGroup}
\title{Group-wise missing value imputation}
\usage{
imputeGroup(x, group = NULL, do.mar = TRUE, do.mnar = TRUE,
  mnar.default = 0, method = "knn", ...)
}
\arguments{
\item{x}{matrix with missing values to impute.}

\item{group}{grouping (column) variable.}

\item{do.mar}{logical; whether to perform MAR imputation.}

\item{do.mnar}{logical; whether to perform MNAR imputation.}

\item{mnar.default}{numerical; default value assigned to MNAR missing values.}

\item{method}{method used for imputation.}

\item{...}{additional arguments passed to the imputation methods.}
}
\value{
a matrix with missing values imputed.
}
\description{
By default MNAR missing values are assigned 0. MAR values are imputed with some of the methods.
}
\examples{
NULL
}
