% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{melt}
\alias{melt}
\title{Custom implementation of reshape2 melt to avoid importing reshape2}
\usage{
melt(x, rows.name = "rowname", cols.name = "colname",
  value.name = "value")
}
\arguments{
\item{x}{matrix to melt.}

\item{rows.name}{name of column containing rownames.}

\item{cols.name}{name of column containing colnames.}

\item{value.name}{name of column containing values.}
}
\value{
a melted data.frame.
}
\description{
Custom implementation of reshape2 melt to avoid importing reshape2
}

