% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Plot a heatmap of a correlation matrix.}
\usage{
plotCorrelation(x, title = "Sample correlation", cluster = FALSE)
}
\arguments{
\item{x}{and object from which an matrix can be obtained.}

\item{title}{title of the plot.}

\item{cluster}{logical; whether to cluster rows/columns.}
}
\description{
Computes the correlation matrix and passes it to plotHeatmap() with
appropriate arguments.
}

