% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotExpression}
\alias{plotExpression}
\title{Plots a heatmap of a matrix}
\usage{
plotExpression(x, cluster = TRUE, scale = FALSE, cpm = FALSE,
  prior.count = 2, log = TRUE)
}
\arguments{
\item{x}{object to plot.}

\item{cluster}{logical specifying whether to apply hierarchichal clustering to the rows.}

\item{scale}{logical specifying whether to scale the rows.}

\item{cpm}{logical specifying}

\item{prior.count}{numeric value passed to cpm().}

\item{log}{logical passed to cpm() to specified whether values should be log-transformed.}
}
\value{
This function returns no value but has the side effect of producing a plot.
}
\description{
Plots a heatmap of a matrix
}

