% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGene}
\alias{plotGene}
\title{Basic gene ploting function}
\usage{
plotGene(symbol, genome, ...)
}
\arguments{
\item{symbol}{symbol of the gene to plot. Will be used to search using BiomartGeneRegionTrack}

\item{genome}{genome assembly (e.g. mm10 or hg38)}

\item{...}{further arguments passed to plotTracks}
}
\value{
nothing but produces a plot as side effect.
}
\description{
Basic gene ploting function
}

