% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,DGEList-method}
\alias{plotHeatmap,EList-method}
\alias{plotHeatmap,ExpressionSet-method}
\alias{plotHeatmap,matrix-method}
\title{Plots a heatmap of a matrix}
\usage{
plotHeatmap(x, row.cluster = TRUE, col.cluster = FALSE, scale = FALSE)

\S4method{plotHeatmap}{matrix}(x, row.cluster = TRUE, col.cluster = FALSE,
  scale = FALSE)

\S4method{plotHeatmap}{ExpressionSet}(x, row.cluster = TRUE,
  col.cluster = FALSE, scale = FALSE)

\S4method{plotHeatmap}{EList}(x, row.cluster = TRUE, col.cluster = FALSE,
  scale = FALSE)

\S4method{plotHeatmap}{DGEList}(x, row.cluster = TRUE, col.cluster = FALSE,
  scale = FALSE)
}
\arguments{
\item{x}{object to plot.}

\item{row.cluster}{logical specifying whether to apply hierarchichal clustering to the rows.}

\item{col.cluster}{logical specifying whether to apply hierarchichal clustering to the columns.}

\item{scale}{logical specifying whether to scale the rows.}
}
\value{
This function returns no value but has the side effect of producing a plot.
}
\description{
Plots a heatmap with row and/or col clustering if specifief.
}
\examples{
NULL
}

