% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairs.R
\name{plotPairs}
\alias{plotPairs}
\title{plotPairs}
\usage{
plotPairs(x)
}
\arguments{
\item{x}{matrix to plot.}
}
\description{
Plots a matrix of pairwise scatter plots (similar to graphics::pairs and GGally::ggpairs).
}
\examples{
NULL
}

