% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoints.R
\docType{methods}
\name{plotPoints}
\alias{plotPoints}
\alias{plotPoints,DGEList-method}
\alias{plotPoints,EList-method}
\alias{plotPoints,ExpressionSet-method}
\alias{plotPoints,matrix-method}
\title{Plot jittered points by group from a matrix.}
\usage{
plotPoints(x, group = NULL, label = NULL, scales = "fixed", ...)

\S4method{plotPoints}{matrix}(x, group = NULL, label = NULL,
  scales = "fixed")

\S4method{plotPoints}{ExpressionSet}(x, group = NULL, label = NULL,
  scales = "fixed", groupCol = NULL)

\S4method{plotPoints}{EList}(x, group = NULL, label = NULL,
  scales = "fixed", groupCol = NULL)

\S4method{plotPoints}{DGEList}(x, group = NULL, label = NULL,
  scales = "fixed", groupCol = NULL)
}
\arguments{
\item{x}{a matrix or an object that can be coherce to a matrix.}

\item{group}{grouping variable (columns).}

\item{label}{custom labels (rows).}

\item{scales}{scales argument passed down to facet_wrap (default: fixed).}

\item{...}{further arguments passed to methods.}

\item{groupCol}{Column name from which extract grouping variable.}
}
\description{
Takes a matrix and produces a dots plot using geom jitter.
}
\details{
Alternatively points are grouped by the groups assigned to the
group argument, which is applied to the columns. Each row is plotted in a
different panel using facet_wrap. The label of the facets is by default
the rownames but it can be passed using the argument label. The scales
of the plots are by default the same but this can be controled with the
scale argument (e.g. changed to 'free_y').
}

