% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPoints.matrix}
\alias{plotPoints.matrix}
\title{Plot jittered points by group from a matrix.}
\usage{
plotPoints.matrix(x, group = NULL, label = NULL, scales = "fixed")
}
\arguments{
\item{x}{a matrix or an object that can be coherce to a matrix.}

\item{group}{grouping variable (columns).}

\item{label}{custom labels (rows).}

\item{scales}{scales argument passed down to facet_wrap (default: fixed).}
}
\description{
This is a simpler alternative to plotPoint() that takes as an argument a matrix.
}

