% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPoints}
\alias{plotPoints}
\title{Plot jittered points by group.}
\usage{
plotPoints(x, selection = NULL, group = NULL, groupCol = "group",
  cpm = FALSE, prior.count = 2, log = TRUE, scales = "fixed")
}
\arguments{
\item{x}{an object from which a matrix can be obtained.}

\item{selection}{character or numeric vector with a selection to plot.}

\item{group}{character vector with grouping variable.}

\item{groupCol}{grouping variable name (to obtain group from x).}

\item{cpm}{whether to compute cpm().}

\item{prior.count}{argument for cpm().}

\item{log}{argument for cpm().}

\item{scales}{argument passed to facet_wrap().}
}
\description{
Plot jittered points by group.
}

