% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRanges.R
\docType{methods}
\name{plotRanges}
\alias{plotRanges}
\alias{plotRanges,GRanges-method}
\alias{plotRanges,IRanges-method}
\title{Plot ranges avoiding overlapping}
\usage{
plotRanges(x, sep = 0.5, ...)

\S4method{plotRanges}{IRanges}(x, sep = 0.5)

\S4method{plotRanges}{GRanges}(x, sep = 0.5, seqnames = NULL)
}
\arguments{
\item{x}{an IRanges object or some object that can be cohereced to IRanges.}

\item{sep}{vertical separation between ranges.}

\item{...}{further arguments passed down to methods.}

\item{seqnames}{sequence name of the ranges.}
}
\description{
This function plots a set of ranges in their locations. Overlapping is
avoided. This function is based on a similar one in the IRanges vignette,
but using ggplot2 graphics instead.
}
\examples{
NULL
}

