% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRanges}
\alias{plotRanges}
\title{Plot ranges avoiding overlapping}
\usage{
plotRanges(x, sep = 0.5)
}
\arguments{
\item{x}{an IRanges object or some object that can be cohereced to IRanges.}

\item{sep}{vertical separation between ranges.}
}
\description{
This function plots a set of ranges in their locations. Overlapping is avoided. This function is based on a similar
one in the IRanges vignette, but using ggplot2 graphics instead.
}

