% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotResult}
\alias{plotResult}
\title{Plot a heatmap of TestResult object from limma package}
\usage{
plotResult(x)
}
\arguments{
\item{x}{TestResult object.}
}
\description{
This plots a heatmap of the TestResult object returned by decideTests. The
columns are the coefficients and the rows are the genes.
}
\examples{
NULL
}
