% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVoom}
\alias{plotVoom}
\title{Plots the mean-variance trend.}
\usage{
plotVoom(x)
}
\arguments{
\item{x}{Output of voom() function.}
}
\description{
This function uses the information stored in the object returned by voom()
when called with save.plot = TRUE.
}
\examples{
NULL
}
