% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{q3tools}
\alias{q3tools}
\alias{q3tools-package}
\title{A bunch of shabby tools for the omics data analyst}
\description{
These package contains some plotting tools at the moment. They help make my
life easier when analyzing omics data. But they are not neccesarily the best
way to do things. Indeed, I expect them to contain many errors, bad choices
and limitations. You have been warned.
}

