% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getEnrichmentResults}
\alias{getEnrichmentResults}
\title{Create a matrix with several objects resulting from enrichment analysis with kegga() and goana()}
\usage{
getEnrichmentResults(..., what = "P.Up", ontology = "BP",
  use.name = FALSE, short.names = FALSE, minlength = 40)
}
\arguments{
\item{...}{results from enrichment}

\item{what}{what to plot: P.Up (default) or P.Down.}

\item{ontology}{for GO results what ontology to use (default BP).}

\item{use.name}{logical; whether to return the term's name or the ids.}

\item{short.names}{whether to use abbreviate() to shorten rownames (handy for long GO descriptions).}

\item{minlength}{minimum length for abbreviate().}
}
\value{
a matrix with p-values from several enrichment analyses.
}
\description{
Create a matrix with several objects resulting from enrichment analysis with kegga() and goana()
}
\examples{
NULL
}
