% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGene}
\alias{plotGene}
\title{Basic gene ploting function}
\usage{
plotGene(symbol, genome, add.ideogram = FALSE, add.data = NULL,
  from = NULL, to = NULL, ...)
}
\arguments{
\item{symbol}{symbol of the gene to plot. Will be used to search using BiomartGeneRegionTrack.}

\item{genome}{genome assembly (e.g. mm10 or hg38).}

\item{add.ideogram}{whether to add an IdeogramTrack (FALSE by default to spead up testing).}

\item{add.data}{list with data to be added as a DataTrack.}

\item{from}{start genomic coordinates.}

\item{to}{end genomic coordinates.}

\item{...}{further arguments passed to plotTracks.}
}
\value{
nothing but produces a plot as side effect.
}
\description{
Plots a gene from a specified assembly using Gviz package. Optionaly plots some data (e.g. read alignments). Enables zooming.
}
\examples{
NULL
}

