% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMds}
\alias{plotMds}
\title{plotMds}
\usage{
plotMds(x, group = NULL)
}
\arguments{
\item{x}{a matrix object.}

\item{group}{grouping variable.}
}
\description{
ggplo2-style MDS plot using plotMDS() in the limma package to get the data.
}
\examples{
NULL
}
