% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPoints}
\alias{plotPoints}
\title{Plot jittered points by group.}
\usage{
plotPoints(x, selection = NULL, groupCol = "group", cpm = TRUE,
  prior.count = 2, log = TRUE)
}
\arguments{
\item{x}{an object from which a matrix can be obtained.}

\item{selection}{character or numeric vector with a selection to plot.}

\item{groupCol}{grouping variable.}

\item{cpm}{whether to compute cpm().}

\item{prior.count}{argument for cpm()}

\item{log}{argument for cpm()}
}
\description{
Plot jittered points by group.
}

