% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVoom}
\alias{plotVoom}
\title{Plots the mean-variance trend as computed by voom().}
\usage{
plotVoom(x)
}
\arguments{
\item{x}{Output of voom() function (with save.plot = TRUE)}
}
\description{
Plots the mean-variance trend as computed by voom().
}

